social_media_api/urls.py
# social_media_api/urls.py
from django.contrib import admin
from django.urls import path, include
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

urlpatterns = [
    path('admin/', admin.site.urls),

    # Auth & accounts
    path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    path("api/accounts/", include("accounts.urls")),
    # Posts app (posts, comments, feed, likes)
    path('api/', include('posts.urls')),

    # Notifications
    path('api/notifications/', include('notifications.urls')),
]
