requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0,<3.16.0
djangorestframework-simplejwt>=5.3.0,<6.0.0
psycopg2-binary>=2.9.9,<3.0
python-dotenv>=1.0.1,<2.0
